tambah paginate

+++ src/pages/sales/MerchantData.jsx
@@ -3,16 +3,32 @@ import {Link,Redirect} from 'react-router-dom'
 import { useSelector } from 'react-redux';
 import axios from '../../config/api'
 
+import Swal from 'sweetalert2'
+
+// paginate
+import Paginator from 'react-hooks-paginator';
+import './MerchantData.css'
+
 export default function MerchantData() {
 
     // dari redux dev tools = 4444 = staff_id
     const staff_id = useSelector(state => state.auth.staff_id)
 
     const [merchant, setMerchant] = useState([])
+    
+    // paginate
+    const pageLimit = 5;
+    const [offset, setOffset] = useState(0);
+    const [currentPage, setCurrentPage] = useState(1);
+    const [currentData, setCurrentData] = useState([]);
 
     useEffect(() => {
         getMerchant();
     }, [])
+   
+    useEffect(() => {
+      setCurrentData(merchant.slice(offset, offset + pageLimit));
+    }, [offset, merchant]);
 
     const getMerchant = () =>{
         console.log(staff_id);
@@ -26,13 +42,23 @@ export default function MerchantData() {
         console.log("id merchant " + id); // 44
         console.log("staff_id" + staff_id); // 10 
 
-        axios.delete(`/merchant/sales/delete/${id}/${staff_id}`).then((res) =>{
-            console.log(res);
-            getMerchant()
-        }).catch(err => console.log(err))
+        Swal.fire({
+            title: 'Yakin?',
+            showCancelButton: true,
+            confirmButtonText: `Delete`
+          }).then((result) => {
+            /* Read more about isConfirmed, isDenied below */
+            if (result.isConfirmed) {
+                axios.delete(`/merchant/sales/delete/${id}/${staff_id}`).then((res) =>{
+                    console.log(res);
+                    getMerchant()
+                }).catch(err => console.log(err))
+              Swal.fire('Deleted!', '', 'success')
+            } 
+          })
     }
 
-    const renderMerchant = merchant.map((mer, index) => {
+    const renderMerchant = currentData.map((mer, index) => {
         console.log(merchant);
         const urlStoreImage = `http://localhost:2020/merchant/read/storeimage`
             return (
@@ -65,6 +91,15 @@ export default function MerchantData() {
             <div className="row">
                 {renderMerchant}
             </div>
+
+            <Paginator
+                totalRecords={merchant.length}
+                pageLimit={pageLimit}
+                pageNeighbours={2}
+                setOffset={setOffset}
+                currentPage={currentPage}
+                setCurrentPage={setCurrentPage}
+            />
         </div>        
     ) : (
         <Redirect to="/login" />
